# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Calculator Project",
    "description": "My User Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/users/{id}": {
      "get": {
        "tags": ["Search User by ID"],
        "summary": "Search user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID number of user to search for",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This service allows you to search for a user with the id number searched"
          }
        }
      }
    },
    "/users/": {
        "post": {
          "tags": ["update old user or add new user"],
          "summary": "update old user or add new user",
          "parameters": [
            {
                "name": "Users",
                "in":"body",
                "description": "update old user or add new user",
                "schema": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        },
                        "country": {
                            "type": "string"
                        }
                    }
                }
            }
          ],
          "responses": {
            "200": {
              "description": "This service allows you to add two numbers together"
            }
          }
        }
      }
  }
}